+++ tests/src/all_suites.adb
@@ -4,6 +4,7 @@ with Result_Tests;
 with Context_Tests;
 with Engine_Tests;
 with Standard_Tests;
+with Composition_Tests;
 
 package body All_Suites is
 
@@ -17,6 +18,7 @@ package body All_Suites is
       Context_Tests.Add_Tests (Result);
       Engine_Tests.Add_Tests (Result);
       Standard_Tests.Add_Tests (Result);
+      Composition_Tests.Add_Tests (Result);
       return Result;
    end Suite;